Restructured code to improve readability ease documentation
