+++ lib/handlers/authentication_handler.dart
