+++ lib/home.dart
@@ -1,13 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_application_1/tabs/booking_view.dart';
-import 'package:flutter_application_1/tabs/booking_view2.dart';
-import 'package:flutter_application_1/tabs/bookings.dart';
-import 'package:flutter_application_1/tabs/profile_settings.dart';
+import 'package:flutter_application_1/tabs/book_tab.dart';
+import 'package:flutter_application_1/tabs/my_bookings_tab.dart';
+import 'package:flutter_application_1/tabs/menu_tab.dart';
 
 /// A tab-based view that acts as the apps top most page.
 ///
 /// This page is navigated to when a user has logged in.
-/// The page has three tabs selecatble by the user: [BookingView], [Bookings] and [Profile].
+/// The page has three tabs selecatble by the user: [BookingView], [MyBookingsTab] and [MenuTab].
 class Home extends StatefulWidget {
   const Home({Key? key}) : super(key: key);
 
@@ -22,9 +21,9 @@ class _HomeState extends State<Home> {
   
   /// The list of tabs/pages selectable using the BottomNavigationBar.
   final List<Widget> _tabs = <Widget>[
-    BookingView2(),
-    const Bookings(),
-    const Profile()
+    BookTab(),
+    const MyBookingsTab(),
+    const MenuTab()
   ];
   
   /// Switches tab to [index].
